+++ src/main/java/com/meti/util/console/Console.java
@@ -1,4 +1,4 @@
-package com.meti.util;
+package com.meti.util.console;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
