+++ src/main/java/com/meti/util/console/ConsoleState.java
@@ -1,4 +1,6 @@
-package com.meti.util;
+package com.meti.util.console;
+
+import com.meti.util.State;
 
 /**
  * @author SirMathhman
