+++ src/main/java/com/meti/util/fxml/Controller.java
@@ -1,4 +1,6 @@
-package com.meti.util;
+package com.meti.util.fxml;
+
+import com.meti.util.console.Console;
 
 /**
  * @author SirMathhman
