+++ src/main/java/com/meti/util/fxml/FXMLBundle.java
@@ -1,4 +1,4 @@
-package com.meti.util;
+package com.meti.util.fxml;
 
 import javafx.scene.Parent;
